// apps/web/react/app/src/generated/api/models/Board.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Board = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    archivedOriginalTitle?: string;
    boardVersion: number;
    createdAt: string;
    id: string;
    ownerUserId: string;
    title: string;
    updatedAt: string;
};